host of bugs: vectorization is borked; inverse was off by a transpose; cannot print ctypes data and expect things to go well

+++ ctest/run.py
@@ -14,7 +14,7 @@ f = interpolate(Expression("x[0]*x[0] + x[1]*x[1]"), space)
 #Diderot stuff:
 #choose types...
 intTy = ct.c_int32
-floatTy = ct.c_float
+floatTy = ct.c_double
 # build json
 jsonFile = "test.json"
 dataFile = "ugg.dill"
@@ -24,7 +24,7 @@ else:
     fb.spaceToJson(space, "test.json", refCellDefault="simplex")
 # build data
 (preFemArgs, femArgs) = fb.passAll(f, intTy, floatTy)
-
+# print(preFemArgs)
 programNameArg = "justTypes"
 nameSpaceArg = "justTypes"
 library = ct.CDLL("./" + programNameArg + ".so")
